requirements.txt
fastapi
uvicorn
python-decouple
email-validator